+++ agent/puppet.rb
@@ -19,25 +19,38 @@ module MCollective
 
       def run(command, options)
         if MCollective::Util.windows?
-          require 'win32/process'
           # If creating the process doesn't outright fail, assume everything
           # was okay. The caller wants to know our exit code, so we'll just use
           # 0 or 1.
           begin
-            Process.create(:command_line => command, :creation_flags => Process::CREATE_NO_WINDOW)
+            ::Process.spawn(command, :new_pgroup => true)
             0
-          rescue #Process::Error => e
+          rescue ::Process::Error => e
             1
           end
         else
-          super
+          # On Unices double fork and exec to run puppet in a disowned child
+          child = fork {
+            grandchild = fork {
+              exec command
+            }
+            if grandchild != nil
+              ::Process.detach(grandchild)
+            end
+          }
+          return 1 if child.nil?
+          ::Process.detach(child)
+          return 0
         end
       end
 
       action "disable" do
         begin
-          msg = @puppet_agent.disable!(request.fetch(:message, "Disabled via MCollective by %s at %s" % [request.caller, Time.now.strftime("%F %R")]))
-          reply[:status] = "Succesfully locked the Puppet agent: %s" % msg
+          request_msg = request.fetch(:message, "Disabled via MCollective by " \
+                                      "%s at %s" % [request.caller,
+                                                    Time.now.strftime("%F %R")])
+          agent_msg = @puppet_agent.disable!(request_msg)
+          reply[:status] = "Succesfully locked the Puppet agent: %s" % agent_msg
         rescue => e
           reply.fail(reply[:status] = "Could not disable Puppet: %s" % e.to_s)
         end
@@ -59,17 +72,23 @@ module MCollective
       action "last_run_summary" do
         summary = @puppet_agent.load_summary
 
-        reply[:type_distribution] = @puppet_agent.managed_resource_type_distribution
+        if request[:parse_log]
+          reply[:logs] = @puppet_agent.last_run_logs
+        else
+          reply[:logs] = {}
+        end
+
+        reply[:type_distribution]     = @puppet_agent.managed_resource_type_distribution
         reply[:out_of_sync_resources] = summary["resources"].fetch("out_of_sync", 0)
-        reply[:failed_resources] = summary["resources"].fetch("failed", 0)
-        reply[:changed_resources] = summary["resources"].fetch("changed", 0)
-        reply[:total_resources] = summary["resources"].fetch("total", 0)
-        reply[:total_time] = summary["time"].fetch("total", 0)
+        reply[:failed_resources]      = summary["resources"].fetch("failed", 0)
+        reply[:changed_resources]     = summary["resources"].fetch("changed", 0)
+        reply[:total_resources]       = summary["resources"].fetch("total", 0)
+        reply[:total_time]            = summary["time"].fetch("total", 0)
         reply[:config_retrieval_time] = summary["time"].fetch("config_retrieval", 0)
-        reply[:lastrun] = Integer(summary["time"].fetch("last_run", 0))
-        reply[:since_lastrun] = Integer(Time.now.to_i - reply[:lastrun])
-        reply[:config_version] = summary["version"].fetch("config", "unknown")
-        reply[:summary] = summary
+        reply[:lastrun]               = Integer(summary["time"].fetch("last_run", 0))
+        reply[:since_lastrun]         = Integer(Time.now.to_i - reply[:lastrun])
+        reply[:config_version]        = summary["version"].fetch("config", "unknown")
+        reply[:summary]               = summary
       end
 
       action "status" do
@@ -79,22 +98,43 @@ module MCollective
       end
 
       action "resource" do
-        allow_managed_resources_management = !!@config.pluginconf.fetch("puppet.resource_allow_managed_resources", "false").match(/^1|true|yes/)
-        resource_types_whitelist = @config.pluginconf.fetch("puppet.resource_type_whitelist", nil)
-        resource_types_blacklist = @config.pluginconf.fetch("puppet.resource_type_blacklist", nil)
-
-        reply.fail!("You cannot specify both puppet.resource_type_whitelist and puppet.resource_type_blacklist in the config file") if resource_types_whitelist && resource_types_blacklist
+        allow_managed_resources_management = \
+          !!@config.pluginconf.fetch("puppet.resource_allow_managed_resources",
+                                     "false").match(/^1|true|yes/)
+        resource_types_whitelist = \
+          @config.pluginconf.fetch("puppet.resource_type_whitelist", nil)
+        resource_types_blacklist = \
+          @config.pluginconf.fetch("puppet.resource_type_blacklist", nil)
+       
+       if resource_types_whitelist && resource_types_blacklist
+          reply.fail!("You cannot specify both puppet.resource_type_whitelist " \
+                      "and puppet.resource_type_blacklist in the config file")
+        end
 
         # if 'none' is specified whitelist nothing
         resource_types_whitelist = "" if resource_types_whitelist == "none"
 
-        # if neither is specified default to whitelisting nothing thus denying everything
-        resource_types_whitelist = "" if resource_types_blacklist.nil? && resource_types_whitelist.nil?
+        # if neither is specified default to whitelisting
+        # nothing thus denying everything
+        if resource_types_blacklist.nil? && resource_types_whitelist.nil?
+          resource_types_whitelist = ""
+        end
+
 
         params = request.data.clone
         params.delete(:process_results)
         type = params.delete(:type).downcase
         resource_name = "%s[%s]" % [type.to_s.capitalize, params[:name]]
+
+        resource_names_whitelist = \
+          @config.pluginconf.fetch("puppet.resource_name_whitelist.%s" % type, nil)
+        resource_names_blacklist = \
+          @config.pluginconf.fetch("puppet.resource_name_blacklist.%s" % type, nil)
+
+        if resource_names_whitelist && resource_names_blacklist
+          reply.fail!("You cannot specify both puppet.resource_name_whitelist.%s " \
+                      "and puppet.resource_name_blacklist.%s in the config file" % [type, type])
+        end
 
         if resource_types_blacklist
           if resource_types_blacklist.split(",").include?(type)
@@ -106,7 +146,18 @@ module MCollective
           end
         end
 
-        if allow_managed_resources_management || !@puppet_agent.managing_resource?(resource_name)
+        if resource_names_blacklist
+          if resource_names_blacklist.split(",").include?(params[:name])
+            reply.fail!("The %s name is listed in the name blacklist" % params[:name])
+          end
+        elsif resource_names_whitelist
+          unless resource_names_whitelist.split(",").include?(params[:name])
+            reply.fail!("The %s name is not listed in the name whitelist" % params[:name])
+          end
+        end
+
+        if allow_managed_resources_management \
+           || !@puppet_agent.managing_resource?(resource_name)
           resource = ::Puppet::Type.type(type).new(params)
           report = ::Puppet::Transaction::Report.new(:mcollective)
           ::Puppet::Util::Log.newdestination(report)
@@ -122,9 +173,12 @@ module MCollective
 
           reply[:changed] = report.resource_statuses[resource_name].changed
 
-          reply.fail!("Failed to apply %s: %s" % [resource_name, reply[:result]]) if report.resource_statuses[resource_name].failed
+          if report.resource_statuses[resource_name].failed
+            reply.fail!("Failed to apply %s: %s" % [resource_name, reply[:result]])
+          end
         else
-          reply.fail!("Puppet is managing the resource '%s', refusing to create conflicting states" % resource_name)
+          reply.fail!("Puppet is managing the resource '%s', " \
+                      "refusing to create conflicting states" % resource_name)
         end
       end
 
@@ -151,7 +205,11 @@ module MCollective
 
         # we can only pass splay arguments if the daemon isn't running :(
         unless @puppet_agent.status[:daemon_present]
-          unless request[:force] == true
+          if request[:force] == true
+            # forcing implies --no-splay
+            args[:splay] = false
+          else
+            # respect splay options
             args[:splay] = request[:splay] if request.include?(:splay)
             args[:splaylimit] = request[:splaylimit] if request.include?(:splaylimit)
 
@@ -174,24 +232,35 @@ module MCollective
         command = [@puppet_command].concat(options).join(" ")
 
         case run_method
-          when :run_in_background
-            Log.debug("Initiating a background puppet agent run using the command: %s" % command)
-            exitcode = run(command, :stdout => :summary, :stderr => :summary, :chomp => true)
+          when :run_in_foreground
+            Log.debug("Initiating a puppet agent run using the command: %s" % command)
+
+            exitcode = run(command, {
+              :stdout => :summary,
+              :stderr => :summary,
+              :chomp => true,
+            })
 
             unless exitcode == 0
-              reply.fail!(reply[:summary] = "Puppet command '%s' had exit code %d, expected 0" % [command, exitcode])
+              reply.fail!(reply[:summary] = "Puppet command '%s' had exit " \
+                                            "code %d, expected 0" \
+                                            % [command, exitcode])
             else
-              reply[:summary] = "Started a background Puppet run using the '%s' command" % command
+              reply[:summary] = "Started a Puppet run using the " \
+                                "'%s' command" % command
             end
 
           when :signal_running_daemon
-            Log.debug("Signaling the running Puppet agent to start an immediate run")
+            Log.debug("Signaling the running Puppet agent " \
+                      "to start an immediate run")
             @puppet_agent.signal_running_daemon
             reply[:summary] = "Signalled the running Puppet Daemon"
 
           else
-            reply.fail!(reply[:summary] = "Do not know how to do puppet runs using method %s" % run_method)
+            reply.fail!(reply[:summary] = "Do not know how to do puppet runs " \
+                                          "using method %s" % run_method)
         end
+        reply[:initiated_at] = Time.now.to_i
       end
     end
   end