+++ util/puppet_agent_mgr/mgr_v3.rb
@@ -1,20 +1,14 @@
 module MCollective
   module Util
-    module PuppetAgentMgr::V3
-      class Manager
-        include PuppetAgentMgr::Common
+    class PuppetAgentMgr
+      class MgrV3 < PuppetAgentMgr
 
-        if Puppet.features.microsoft_windows?
-          require '%s/windows.rb' % File.dirname(__FILE__)
-          include Windows
-        else
-          require '%s/unix.rb' % File.dirname(__FILE__)
-          include Unix
-        end
-
-        def initialize(configfile=nil, puppet_service='puppet', testing=false)
-          @puppet_service = puppet_service
+        def initialize(configfile   = nil,
+                       service_name = 'puppet',
+                       testing      = false)
+          @puppet_service = service_name
           unless testing || Puppet.settings.app_defaults_initialized?
+
             require 'puppet/util/run_mode'
             Puppet.settings.preferred_run_mode = :agent
 
@@ -22,7 +16,14 @@ module MCollective
             (args << "--config=%s" % configfile) if configfile
 
             Puppet.settings.initialize_global_settings(args)
-            Puppet.settings.initialize_app_defaults(Puppet::Settings.app_defaults_for_run_mode(Puppet.run_mode))
+            Puppet.settings.initialize_app_defaults(
+              Puppet::Settings.app_defaults_for_run_mode(Puppet.run_mode))
+            # This check is to keep backwards compatibility
+            # with Puppet versions < 3.5.0
+            if Puppet.respond_to?(:base_context) \
+                && Puppet.respond_to?(:push_context)
+              Puppet.push_context(Puppet.base_context(Puppet.settings))
+            end
           end
         end
 
@@ -32,34 +33,14 @@ module MCollective
           File.unlink(Puppet[:agent_disabled_lockfile])
         end
 
-        # disable the puppet agent, on version 2.x the message is ignored
+        # disable the puppet agent
         def disable!(msg=nil)
           raise "Already disabled" unless enabled?
-
-          msg = "Disabled using the Ruby API at %s" % Time.now.strftime("%c") unless msg
-
+          msg ||= "Disabled using the Ruby API at %s" % Time.now.strftime("%c")
           atomic_file(Puppet[:agent_disabled_lockfile]) do |f|
             f.print(JSON.dump(:disabled_message => msg))
           end
-
           msg
-        end
-
-        # all the managed resources
-        def managed_resources
-          # need to add some caching here based on mtime of the resources file
-          return [] unless File.exist?(Puppet[:resourcefile])
-
-          File.readlines(Puppet[:resourcefile]).map do |resource|
-            resource.chomp
-          end
-        end
-
-        # epoch time when the last catalog was applied
-        def lastrun
-          summary = load_summary
-
-          Integer(summary["time"].fetch("last_run", 0))
         end
 
         # the current lock message, always "" on 2.0
@@ -75,6 +56,17 @@ module MCollective
         # is the agent disabled
         def disabled?
           File.exist?(Puppet[:agent_disabled_lockfile])
+        end
+
+        private
+
+        def platform_applying?
+          return false unless File.exist?(Puppet[:agent_catalog_run_lockfile])
+          if File::Stat.new(Puppet[:agent_catalog_run_lockfile]).size > 0
+            return has_process_for_pid?(
+                      File.read(Puppet[:agent_catalog_run_lockfile]))
+          end
+          return false
         end
       end
     end