+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 |Release Version|Date|Description|Ticket|
 |---------------|----|-----------|------|
+|1.7.2|2014/04/25|Remove implicit requirement on puppet 3.5.x|MCOP-25|
+|1.7.2|2014/04/24|Fix puppet-agent application backtrace when no results are returned|MCOP-26|
+|1.7.1|2014/04/23|MCollective service plugin not working with puppet 3.5.x|MCOP-23|
+|1.7.0|2014/02/20|Fully-qualify uses of Process to avoid clashes with process agent|#13|
+|1.7.0|2014/02/20|Fix --no-noop and --no-splay under MCollective 2.3.x and 2.4.x|MCOP-5|
+|1.7.0|2014/01/10|Change method of running puppet agent to double-fork a foreground run|MCO-31|
+|1.6.2|2013/12/04|Change noop to no-op for front-end text|MCO-28|
+|1.6.1|2013/10/15|Made --force option correctly imply --no-splay|22860|
 |1.6.0|2013/05/08|Increase the DDL timeout to better handle slower servers where puppet startup is slow|20618 |
 |1.6.0|2013/05/07|Support controlling Puppet on Windows|19541|
 |1.5.1|2013/03/01|Add a --rerun option to the runall command that loops over the nodes forever|19541|